requirements.txt
django==3.2.13
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
requests==2.26.0
pytz==2020.1
sqlparse==0.3.1
django-colorfield==0.7.0
django_cleanup==6.0.0
reportlab==3.6.9
drf-extra-fields==3.4.0